requirements.txt
Flask==0.10
Flask-Login==0.2.4
alembic>=0.6.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.3
Flask-Script>=0.6.3